+++ src/app/sub/submission/[id]/page.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { NextPage } from "next";
 import SubmissionModule from "@/modules/SubmissionModule";
 import { getUserData } from "@/hooks/user";
-import { notFound, redirect } from "next/navigation";
-import { getSubmissionData, getSubmissionItemData } from "@/hooks/submission";
+import { redirect } from "next/navigation";
+import { getSubmissionData, getSubmissionItemData } from "@/hooks/sub";
 import NotFound from "@/app/not-found";
 
 const page: NextPage<{
